requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
requests>=2.31.0
PyPDF2>=3.0.1
python-docx>=1.1.0
pinecone-client>=2.2.4
openai>=1.6.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-pinecone>=0.0.3
langchain-community>=0.0.10
numpy>=1.24.3
pandas>=2.0.3
python-dotenv>=1.0.1
psutil>=5.9.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0